+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # Changelog for extensible-effects-concurrent
 
+## 0.5.0.0
+
+- Switch to `extensible-effects` version `3.1.0.0`
+- Bump to stackage LTS-12.9
+- Add `Control.Eff.Log.MessageFactory`
+- Add `Control.Eff.Log.Syslog`
+
 ## 0.4.0.0
 
 - Switch to `extensible-effects` version `3.0.0.0`